Moar abstraction!!!1!
fml
