+++ WhatsAppApi/WhatsApp.cs
@@ -20,49 +20,11 @@ namespace WhatsAppApi
     /// <summary>
     /// Main api interface
     /// </summary>
-    public class WhatsApp : WhatsAppBase
+    public class WhatsApp : WhatsSendBase
     {
         public WhatsApp(string phoneNum, string imei, string nick, bool debug = false, bool hidden = false)
         {
             this._constructBase(phoneNum, imei, nick, debug, hidden);
-        }
-
-        public void Login(byte[] nextChallenge = null)
-        {
-            //reset stuff
-            this.reader.Key = null;
-            this.BinWriter.Key = null;
-            this._challengeBytes = null;
-
-            if (nextChallenge != null)
-            {
-                this._challengeBytes = nextChallenge;
-            }
-
-            string resource = string.Format(@"{0}-{1}-{2}",
-                WhatsConstants.Device,
-                WhatsConstants.WhatsAppVer,
-                WhatsConstants.WhatsPort);
-            var data = this.BinWriter.StartStream(WhatsConstants.WhatsAppServer, resource);
-            var feat = this.addFeatures();
-            var auth = this.addAuth();
-            this.SendData(data);
-            this.SendData(this.BinWriter.Write(feat, false));
-            this.SendData(this.BinWriter.Write(auth, false));
-
-            this.pollMessage();//stream start
-            this.pollMessage();//features
-            this.pollMessage();//challenge or success
-
-            if (this.loginStatus != CONNECTION_STATUS.LOGGEDIN)
-            {
-                //oneshot failed
-                ProtocolTreeNode authResp = this.addAuthResponse();
-                this.SendData(this.BinWriter.Write(authResp, false));
-                this.pollMessage();
-            }
-
-            this.SendAvailableForChat(this.name, this.hidden);
         }
 
         public void Message(string to, string txt)
@@ -237,7 +199,7 @@ namespace WhatsAppApi
             }
         }
 
-        private WaUploadResponse UploadFile(string b64hash, string type, long size, string path, string to, string contenttype)
+        protected WaUploadResponse UploadFile(string b64hash, string type, long size, string path, string to, string contenttype)
         {
             ProtocolTreeNode media = new ProtocolTreeNode("media", new KeyValue[] {
                 new KeyValue("hash", b64hash),
@@ -356,487 +318,6 @@ namespace WhatsAppApi
             return null;
         }
 
-        public void PollMessages(bool autoReceipt = true)
-        {
-            while (pollMessage(autoReceipt)) ;
-        }
-
-        public bool pollMessage(bool autoReceipt = true)
-        {
-            if (this.loginStatus == CONNECTION_STATUS.CONNECTED || this.loginStatus == CONNECTION_STATUS.LOGGEDIN)
-            {
-                byte[] nodeData;
-                try
-                {
-                    nodeData = this.whatsNetwork.ReadNextNode();
-                    if (nodeData != null)
-                    {
-                        return this.processInboundData(nodeData, autoReceipt);
-                    }
-                }
-                catch (ConnectionException)
-                {
-                    this.Disconnect();
-                }
-            }
-            return false;
-        }
-
-        protected ProtocolTreeNode addFeatures()
-        {
-            return new ProtocolTreeNode("stream:features", null);
-        }
-
-        protected ProtocolTreeNode addAuth()
-        {
-            List<KeyValue> attr = new List<KeyValue>(new KeyValue[] {
-                new KeyValue("mechanism", Helper.KeyStream.AuthMethod),
-                new KeyValue("user", this.phoneNumber)});
-            if (this.hidden)
-            {
-                attr.Add(new KeyValue("passive", "true"));
-            }
-            var node = new ProtocolTreeNode("auth", attr.ToArray(), null, this.getAuthBlob());
-            return node;
-        }
-
-        protected byte[] getAuthBlob()
-        {
-            byte[] data = null;
-            if (this._challengeBytes != null)
-            {
-                byte[][] keys = KeyStream.GenerateKeys(this.encryptPassword(), this._challengeBytes);
-
-                this.reader.Key = new KeyStream(keys[2], keys[3]);
-
-                this.outputKey = new KeyStream(keys[0], keys[1]);
-
-                PhoneNumber pn = new PhoneNumber(this.phoneNumber);
-
-                List<byte> b = new List<byte>();
-                b.AddRange(new byte[] { 0, 0, 0, 0 });
-                b.AddRange(WhatsApp.SYSEncoding.GetBytes(this.phoneNumber));
-                b.AddRange(this._challengeBytes);
-                b.AddRange(WhatsApp.SYSEncoding.GetBytes(Helper.Func.GetNowUnixTimestamp().ToString()));
-                b.AddRange(WhatsApp.SYSEncoding.GetBytes(WhatsConstants.UserAgent));
-                b.AddRange(WhatsApp.SYSEncoding.GetBytes(String.Format(" MccMnc/{0}001", pn.MCC)));
-                data = b.ToArray();
-
-                this._challengeBytes = null;
-
-                this.outputKey.EncodeMessage(data, 0, 4, data.Length - 4);
-
-                this.BinWriter.Key = this.outputKey;
-            }
-
-            return data;
-        }
-
-        protected ProtocolTreeNode addAuthResponse()
-        {
-            if (this._challengeBytes != null)
-            {
-                byte[][] keys = KeyStream.GenerateKeys(this.encryptPassword(), this._challengeBytes);
-
-                this.reader.Key = new KeyStream(keys[2], keys[3]);
-                this.BinWriter.Key = new KeyStream(keys[0], keys[1]);
-
-                List<byte> b = new List<byte>();
-                b.AddRange(new byte[] { 0, 0, 0, 0 });
-                b.AddRange(WhatsApp.SYSEncoding.GetBytes(this.phoneNumber));
-                b.AddRange(this._challengeBytes);
-
-
-                byte[] data = b.ToArray();
-                this.BinWriter.Key.EncodeMessage(data, 0, 4, data.Length - 4);
-                var node = new ProtocolTreeNode("response",
-                    new KeyValue[] { new KeyValue("xmlns", "urn:ietf:params:xml:ns:xmpp-sasl") },
-                    data);
-
-                return node;
-            }
-            throw new Exception("Auth response error");
-        }
-
-        protected void processChallenge(ProtocolTreeNode node)
-        {
-            _challengeBytes = node.data;
-        }
-        
-        protected bool processInboundData(byte[] msgdata, bool autoReceipt = true)
-        {
-            try
-            {
-                ProtocolTreeNode node = this.reader.nextTree(msgdata);
-                if(node != null)
-                {
-                    if (ProtocolTreeNode.TagEquals(node, "challenge"))
-                    {
-                        this.processChallenge(node);
-                    }
-                    else if (ProtocolTreeNode.TagEquals(node, "success"))
-                    {
-                        this.loginStatus = CONNECTION_STATUS.LOGGEDIN;
-                        this.accountinfo = new AccountInfo(node.GetAttribute("status"),
-                                                           node.GetAttribute("kind"),
-                                                           node.GetAttribute("creation"),
-                                                           node.GetAttribute("expiration"));
-                        this.fireOnLoginSuccess(this.phoneNumber, node.GetData());
-                    }
-                    else if (ProtocolTreeNode.TagEquals(node, "failure"))
-                    {
-                        this.loginStatus = CONNECTION_STATUS.UNAUTHORIZED;
-                        this.fireOnLoginFailed(node.children.FirstOrDefault().tag);
-                    }
-                    
-                    if (ProtocolTreeNode.TagEquals(node, "receipt"))
-                    {
-                        string from = node.GetAttribute("from");
-                        string id = node.GetAttribute("id");
-                        string type = node.GetAttribute("type") ?? "delivery";
-                        switch (type)
-                        {
-                            case "delivery":
-                                //delivered to target
-                                this.fireOnGetMessageReceivedClient(from, id);
-                                break;
-                            case "read":
-                                //read by target
-                                //todo
-                                break;
-                            case "played":
-                                //played by target
-                                //todo
-                                break;
-                        }
-
-                        //send ack
-                        SendNotificationAck(node, type);
-                    }
-
-                    if (ProtocolTreeNode.TagEquals(node, "message"))
-                    {
-                        this.handleMessage(node, autoReceipt);
-                    }
-
-
-                    if (ProtocolTreeNode.TagEquals(node, "iq"))
-                    {
-                        this.handleIq(node);
-                    }
-
-                    if (ProtocolTreeNode.TagEquals(node, "stream:error"))
-                    {
-                        var textNode = node.GetChild("text");
-                        if (textNode != null)
-                        {
-                            string content = WhatsApp.SYSEncoding.GetString(textNode.GetData());
-                            Console.WriteLine("Error : " + content);
-                        }
-                        this.Disconnect();
-                    }
-
-                    if (ProtocolTreeNode.TagEquals(node, "presence"))
-                    {
-                        //presence node
-                        this.fireOnGetPresence(node.GetAttribute("from"), node.GetAttribute("type"));
-                    }
-
-                    if (node.tag == "ib")
-                    {
-                        foreach (ProtocolTreeNode child in node.children)
-                        {
-                            switch (child.tag)
-                            {
-                                case "dirty":
-                                    this.SendClearDirty(child.GetAttribute("type"));
-                                    break;
-                                case "offline":
-                                    //this.SendQrSync(null);
-                                    break;
-                                default:
-                                    throw new NotImplementedException(node.NodeString());
-                            }
-                        }
-                    }
-
-                    if (node.tag == "chatstate")
-                    {
-                        string state = node.children.FirstOrDefault().tag;
-                        switch (state)
-                        {
-                            case "composing":
-                                this.fireOnGetTyping(node.GetAttribute("from"));
-                                break;
-                            case "paused":
-                                this.fireOnGetPaused(node.GetAttribute("from"));
-                                break;
-                            default:
-                                throw new NotImplementedException(node.NodeString());
-                        }
-                    }
-
-                    if (node.tag == "ack")
-                    {
-                        string cls = node.GetAttribute("class");
-                        if (cls == "message")
-                        {
-                            //server receipt
-                            this.fireOnGetMessageReceivedServer(node.GetAttribute("from"), node.GetAttribute("id"));
-                        }
-                    }
-
-                    if (node.tag == "notification")
-                    {
-                        this.handleNotification(node);
-                    }
-
-                    return true;
-                }
-            }
-            catch (Exception e)
-            {
-                throw e;
-            }
-            return false;
-        }
-
-        protected void handleMessage(ProtocolTreeNode node, bool autoReceipt)
-        {
-            if (!string.IsNullOrEmpty(node.GetAttribute("notify")))
-            {
-                string name = node.GetAttribute("notify");
-                this.fireOnGetContactName(node.GetAttribute("from"), name);
-            }
-            if (node.GetAttribute("type") == "error")
-            {
-                throw new NotImplementedException(node.NodeString());
-            }
-            if (node.GetChild("body") != null)
-            {
-                //text message
-                this.fireOnGetMessage(node, node.GetAttribute("from"), node.GetAttribute("id"), node.GetAttribute("notify"), System.Text.Encoding.UTF8.GetString(node.GetChild("body").GetData()), autoReceipt);
-                if (autoReceipt)
-                {
-                    this.sendMessageReceived(node);
-                }
-            }
-            if (node.GetChild("media") != null)
-            {
-                ProtocolTreeNode media = node.GetChild("media");
-                //media message
-
-                //define variables in switch
-                string file, url, from, id;
-                int size;
-                byte[] preview, dat;
-                id = node.GetAttribute("id");
-                from = node.GetAttribute("from");
-                switch (media.GetAttribute("type"))
-                {
-                    case "image":
-                        url = media.GetAttribute("url");
-                        file = media.GetAttribute("file");
-                        size = Int32.Parse(media.GetAttribute("size"));
-                        preview = media.GetData();
-                        this.fireOnGetMessageImage(from, id, file, size, url, preview);
-                        break;
-                    case "audio":
-                        file = media.GetAttribute("file");
-                        size = Int32.Parse(media.GetAttribute("size"));
-                        url = media.GetAttribute("url");
-                        preview = media.GetData();
-                        this.fireOnGetMessageAudio(from, id, file, size, url, preview);
-                        break;
-                    case "video":
-                        file = media.GetAttribute("file");
-                        size = Int32.Parse(media.GetAttribute("size"));
-                        url = media.GetAttribute("url");
-                        preview = media.GetData();
-                        this.fireOnGetMessageVideo(from, id, file, size, url, preview);
-                        break;
-                    case "location":
-                        double lon = double.Parse(media.GetAttribute("longitude"), System.Globalization.CultureInfo.InvariantCulture);
-                        double lat = double.Parse(media.GetAttribute("latitude"), System.Globalization.CultureInfo.InvariantCulture);
-                        preview = media.GetData();
-                        name = media.GetAttribute("name");
-                        url = media.GetAttribute("url");
-                        this.fireOnGetMessageLocation(from, id, lon, lat, url, name, preview);
-                        break;
-                    case "vcard":
-                        ProtocolTreeNode vcard = media.GetChild("vcard");
-                        name = vcard.GetAttribute("name");
-                        dat = vcard.GetData();
-                        this.fireOnGetMessageVcard(from, id, name, dat);
-                        break;
-                }
-                this.sendMessageReceived(node);
-            }
-        }
-
-        protected void handleIq(ProtocolTreeNode node)
-        {
-            if (node.GetAttribute("type") == "error")
-            {
-                this.fireOnError(node.GetAttribute("id"), node.GetAttribute("from"), Int32.Parse(node.GetChild("error").GetAttribute("code")), node.GetChild("error").GetAttribute("text"));
-            }
-            if (node.GetChild("sync") != null)
-            {
-                //sync result
-                ProtocolTreeNode sync = node.GetChild("sync");
-                ProtocolTreeNode existing = sync.GetChild("in");
-                ProtocolTreeNode nonexisting = sync.GetChild("out");
-                //process existing first
-                Dictionary<string, string> existingUsers = new Dictionary<string, string>();
-                foreach (ProtocolTreeNode child in existing.GetAllChildren())
-                {
-                    existingUsers.Add(System.Text.Encoding.UTF8.GetString(child.GetData()), child.GetAttribute("jid"));
-                }
-                //now process failed numbers
-                List<string> failedNumbers = new List<string>();
-                foreach (ProtocolTreeNode child in nonexisting.GetAllChildren())
-                {
-                    failedNumbers.Add(System.Text.Encoding.UTF8.GetString(child.GetData()));
-                }
-                int index = 0;
-                Int32.TryParse(sync.GetAttribute("index"), out index);
-                this.fireOnGetSyncResult(index, sync.GetAttribute("sid"), existingUsers, failedNumbers.ToArray());
-            }
-            if (node.GetAttribute("type").Equals("result", StringComparison.OrdinalIgnoreCase)
-                && node.children.FirstOrDefault().tag == "query"
-                && node.children.FirstOrDefault().GetAttribute("xmlns") == "jabber:iq:last"
-            )
-            {
-                //last seen
-                DateTime lastSeen = DateTime.Now.AddSeconds(double.Parse(node.children.FirstOrDefault().GetAttribute("seconds")) * -1);
-                this.fireOnGetLastSeen(node.GetAttribute("from"), lastSeen);
-            }
-            if (node.GetAttribute("type").Equals("result", StringComparison.OrdinalIgnoreCase)
-                && (ProtocolTreeNode.TagEquals(node.children.FirstOrDefault(), "media") || ProtocolTreeNode.TagEquals(node.children.FirstOrDefault(), "duplicate"))
-                )
-            {
-                //media upload
-                this.uploadResponse = node;
-            }
-            if (node.GetAttribute("type").Equals("result", StringComparison.OrdinalIgnoreCase)
-                && ProtocolTreeNode.TagEquals(node.children.FirstOrDefault(), "picture")
-                )
-            {
-                //profile picture
-                string from = node.GetAttribute("from");
-                string id = node.GetChild("picture").GetAttribute("id");
-                byte[] dat = node.GetChild("picture").GetData();
-                string type = node.GetChild("picture").GetAttribute("type");
-                if (type == "preview")
-                {
-                    this.fireOnGetPhotoPreview(from, id, dat);
-                }
-                else
-                {
-                    this.fireOnGetPhoto(from, id, dat);
-                }
-            }
-            if (node.GetAttribute("type").Equals("get", StringComparison.OrdinalIgnoreCase)
-                && ProtocolTreeNode.TagEquals(node.children.FirstOrDefault(), "ping"))
-            {
-                this.SendPong(node.GetAttribute("id"));
-            }
-            if (node.GetAttribute("type").Equals("result", StringComparison.OrdinalIgnoreCase)
-                && ProtocolTreeNode.TagEquals(node.children.FirstOrDefault(), "group"))
-            {
-                //group(s) info
-                List<WaGroupInfo> groups = new List<WaGroupInfo>();
-                foreach (ProtocolTreeNode group in node.children)
-                {
-                    groups.Add(new WaGroupInfo(
-                        group.GetAttribute("id"),
-                        group.GetAttribute("owner"),
-                        long.Parse(group.GetAttribute("creation")),
-                        group.GetAttribute("subject"),
-                        long.Parse(group.GetAttribute("s_t")),
-                        group.GetAttribute("s_o")
-                        ));
-                }
-                this.fireOnGetGroups(groups.ToArray());
-            }
-            if (node.GetAttribute("type").Equals("result", StringComparison.OrdinalIgnoreCase)
-                && ProtocolTreeNode.TagEquals(node.children.FirstOrDefault(), "participant"))
-            {
-                //group participants
-                List<string> participants = new List<string>();
-                foreach (ProtocolTreeNode part in node.GetAllChildren())
-                {
-                    if (part.tag == "participant" && !string.IsNullOrEmpty(part.GetAttribute("jid")))
-                    {
-                        participants.Add(part.GetAttribute("jid"));
-                    }
-                }
-                this.fireOnGetGroupParticipants(node.GetAttribute("from"), participants.ToArray());
-            }
-            if (node.GetAttribute("type") == "result" && node.GetChild("status") != null)
-            {
-                foreach (ProtocolTreeNode status in node.GetChild("status").GetAllChildren())
-                {
-                    this.fireOnGetStatus(status.GetAttribute("jid"),
-                        "result",
-                        null,
-                        WhatsApp.SYSEncoding.GetString(status.GetData()));
-                }
-            }
-        }
-
-        protected void handleNotification(ProtocolTreeNode node)
-        {
-            if (!String.IsNullOrEmpty(node.GetAttribute("notify")))
-            {
-                this.fireOnGetContactName(node.GetAttribute("from"), node.GetAttribute("notify"));
-            }
-            string type = node.GetAttribute("type");
-            switch (type)
-            {
-                case "picture":
-                    ProtocolTreeNode child = node.children.FirstOrDefault();
-                    this.fireOnNotificationPicture(child.tag, child.GetAttribute("jid"), child.GetAttribute("id"));
-                    break;
-                case "status":
-                    ProtocolTreeNode child2 = node.children.FirstOrDefault();
-                    this.fireOnGetStatus(node.GetAttribute("from"), child2.tag, node.GetAttribute("notify"), System.Text.Encoding.UTF8.GetString(child2.GetData()));
-                    break;
-                default:
-                    throw new NotImplementedException(node.NodeString());
-            }
-            this.SendNotificationAck(node);
-        }
-
-        private void SendNotificationAck(ProtocolTreeNode node, string type = null)
-        {
-            string from = node.GetAttribute("from");
-            string to = node.GetAttribute("to");
-            string participant = node.GetAttribute("participant");
-            string id = node.GetAttribute("id");
-            if (type == null)
-            {
-                type = node.GetAttribute("type");
-            }
-            List<KeyValue> attributes = new List<KeyValue>();
-            if(!string.IsNullOrEmpty(to))
-            {
-                attributes.Add(new KeyValue("from", to));
-            }
-            if(!string.IsNullOrEmpty(participant))
-            {
-                attributes.Add(new KeyValue("participant", participant));
-            }
-            attributes.AddRange(new [] {
-                new KeyValue("to", from),
-                new KeyValue("class", node.tag),
-                new KeyValue("id", id),
-                new KeyValue("type", type)
-            });
-            ProtocolTreeNode sendNode = new ProtocolTreeNode("ack", attributes.ToArray());
-            this.SendNode(sendNode);
-        }
-
         protected void SendQrSync(byte[] qrkey, byte[] token = null)
         {
             string id = TicketCounter.MakeId("qrsync_");
@@ -853,13 +334,7 @@ namespace WhatsAppApi
             }, children.ToArray());
             this.SendNode(node);
         }
-
-        public void sendMessageReceived(ProtocolTreeNode msg, string response = "received")
-        {
-            FMessage tmpMessage = new FMessage(new FMessage.FMessageIdentifierKey(msg.GetAttribute("from"), true, msg.GetAttribute("id")));
-            this.SendMessageReceived(tmpMessage, response);
-        }
-
+        
         public void SendActive()
         {
             var node = new ProtocolTreeNode("presence", new[] { new KeyValue("type", "active") });
@@ -872,41 +347,10 @@ namespace WhatsAppApi
             this.SendVerbParticipants(gjid, participants, id, "add");
         }
 
-        public void SendAvailableForChat(string nickName, bool isHidden = false)
-        {
-            var node = new ProtocolTreeNode("presence", new[] { new KeyValue("name", nickName) });
-            this.SendNode(node);
-        }
-
         public void SendUnavailable()
         {
             var node = new ProtocolTreeNode("presence", new[] { new KeyValue("type", "unavailable") });
             this.SendNode(node);
-        }
-
-        public void SendClearDirty(IEnumerable<string> categoryNames)
-        {
-            string id = TicketCounter.MakeId("clean_dirty_");
-            List<ProtocolTreeNode> children = new List<ProtocolTreeNode>();
-            foreach (string category in categoryNames)
-            {
-                ProtocolTreeNode cat = new ProtocolTreeNode("clean", new[] { new KeyValue("type", category) });
-                children.Add(cat);
-            }
-            var node = new ProtocolTreeNode("iq",
-                                            new[]
-                                                {
-                                                    new KeyValue("id", id), 
-                                                    new KeyValue("type", "set"),
-                                                    new KeyValue("to", "s.whatsapp.net"),
-                                                    new KeyValue("xmlns", "urn:xmpp:whatsapp:dirty")
-                                                }, children);
-            this.SendNode(node);
-        }
-
-        public void SendClearDirty(string category)
-        {
-            this.SendClearDirty(new string[] { category });
         }
 
         public void SendClientConfig(string platform, string lg, string lc)
@@ -1000,11 +444,6 @@ namespace WhatsAppApi
             var child = new ProtocolTreeNode("query", new[] { new KeyValue("xmlns", "jabber:iq:roster") }, new ProtocolTreeNode[] { innerChild });
             var node = new ProtocolTreeNode("iq", new[] { new KeyValue("type", "set"), new KeyValue("id", v) }, new ProtocolTreeNode[] { child });
             this.SendNode(node);
-        }
-
-        protected void SendDeliveredReceiptAck(string to, string id)
-        {
-            this.SendReceiptAck(to, id, "delivered");
         }
 
         public void SendEndGroupChat(string gjid)
@@ -1199,26 +638,9 @@ namespace WhatsAppApi
             }
         }
 
-        protected void SendMessageReceived(FMessage message, string response)
-        {
-            ProtocolTreeNode node = new ProtocolTreeNode("receipt", new[] {
-                new KeyValue("to", message.identifier_key.remote_jid),
-                new KeyValue("id", message.identifier_key.id)
-            });
-
-            this.SendNode(node);
-        }
-
         public void SendNop()
         {
             this.SendNode(null);
-        }
-
-        protected void SendNotificationReceived(string jid, string id)
-        {
-            var child = new ProtocolTreeNode("received", new[] { new KeyValue("xmlns", "urn:xmpp:receipts") });
-            var node = new ProtocolTreeNode("message", new[] { new KeyValue("to", jid), new KeyValue("type", "notification"), new KeyValue("id", id) }, child);
-            this.SendNode(node);
         }
 
         public void SendPaused(string to)
@@ -1231,12 +653,6 @@ namespace WhatsAppApi
             string id = TicketCounter.MakeId("ping_");
             var child = new ProtocolTreeNode("ping", new[] { new KeyValue("xmlns", "w:p") });
             var node = new ProtocolTreeNode("iq", new[] { new KeyValue("id", id), new KeyValue("type", "get") }, child);
-            this.SendNode(node);
-        }
-
-        protected void SendPong(string id)
-        {
-            var node = new ProtocolTreeNode("iq", new[] { new KeyValue("type", "result"), new KeyValue("to", WhatsConstants.WhatsAppRealm), new KeyValue("id", id) });
             this.SendNode(node);
         }
 
@@ -1406,18 +822,6 @@ namespace WhatsAppApi
                     new KeyValue("mute", string.Format(System.Globalization.CultureInfo.InvariantCulture, "{0}", new object[] { (!@group.MuteExpiry.HasValue || (@group.MuteExpiry.Value <= now)) ? 0 : ((int) (@group.MuteExpiry.Value - now).TotalSeconds) })) })).ToArray<ProtocolTreeNode>();
             }
             return nodeArray;
-        }
-
-        private void SendReceiptAck(string to, string id, string receiptType)
-        {
-            var tmpChild = new ProtocolTreeNode("ack", new[] { new KeyValue("xmlns", "urn:xmpp:receipts"), new KeyValue("type", receiptType) });
-            var resultNode = new ProtocolTreeNode("message", new[]
-                                                             {
-                                                                 new KeyValue("to", to),
-                                                                 new KeyValue("type", "chat"),
-                                                                 new KeyValue("id", id)
-                                                             }, tmpChild);
-            this.SendNode(resultNode);
         }
 
         protected static ProtocolTreeNode GetMessageNode(FMessage message, ProtocolTreeNode pNode, bool hidden = false)
